+++ datalayer/futures/src/main/java/com/waben/stock/datalayer/futures/service/FuturesContractService.java
@@ -26,6 +26,7 @@ import org.springframework.stereotype.Service;
 import com.waben.stock.datalayer.futures.entity.FuturesContract;
 import com.waben.stock.datalayer.futures.entity.FuturesContractTerm;
 import com.waben.stock.datalayer.futures.entity.FuturesExchange;
+import com.waben.stock.datalayer.futures.entity.FuturesOrder;
 import com.waben.stock.datalayer.futures.repository.DynamicQuerySqlDao;
 import com.waben.stock.datalayer.futures.repository.FuturesContractDao;
 import com.waben.stock.datalayer.futures.repository.impl.MethodDesc;
@@ -50,6 +51,11 @@ public class FuturesContractService {
 
 	@Autowired
 	private FuturesContractDao futuresContractDao;
+	
+	@Autowired
+	private FuturesOrderService orderService;
+	
+	
 
 	public Page<FuturesContract> pagesContract(final FuturesContractQuery query) {
 		Pageable pageable = new PageRequest(query.getPage(), query.getSize());
@@ -83,16 +89,11 @@ public class FuturesContractService {
 			public Predicate toPredicate(Root<FuturesContract> root, CriteriaQuery<?> criteriaQuery,
 					CriteriaBuilder criteriaBuilder) {
 				List<Predicate> predicateList = new ArrayList<Predicate>();
-				// Join<FuturesExchange, FuturesContract> parentJoin =
-				// root.join("exchange", JoinType.LEFT);
 				Join<FuturesContract, FuturesExchange> join = root.join("exchange", JoinType.LEFT);
-				if (query.getExchangcode() != null && !"".equals(query.getExchangcode())) {
-					predicateList.add(criteriaBuilder.equal(join.get("code").as(String.class), query.getExchangcode()));
+				if(query.getExchangcode() != null && !"".equals(query.getExchangcode())){
+					predicateList.add(criteriaBuilder.or(criteriaBuilder.like(join.get("code").as(String.class), query.getExchangcode()+"%"),criteriaBuilder.like(join.get("name").as(String.class), query.getExchangcode()+"%")));
 				}
-				if (query.getExchangename() != null && !"".equals(query.getExchangename())) {
-					predicateList
-							.add(criteriaBuilder.equal(join.get("name").as(String.class), query.getExchangename()));
-				}
+				
 
 				if (query.getSymbol() != null && !"".equals(query.getSymbol())) {
 					predicateList.add(criteriaBuilder.equal(root.get("symbol").as(String.class), query.getSymbol()));
@@ -226,9 +227,17 @@ public class FuturesContractService {
 	public FuturesContract findByContractId(Long contractId) {
 		return futuresContractDao.retrieve(contractId);
 	}
+	
+	public String deleteExchange(Long id){
+		List<Long> contractId = new ArrayList<Long>();
+		contractId.add(id);
+		List<FuturesOrder> list = orderService.findByContractId(contractId);
+		if(list.size()>0){
+			return "该合约正在被订单使用，请不要删除";
+		}
 
-	public void deleteExchange(Long id) {
 		futuresContractDao.delete(id);
+		return "删除成功";
 	}
 
 	public List<FuturesContractTerm> findByListContractId(Long contractId) {
